backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]==1.7.4
python-multipart
pydantic[email]
bcrypt==4.0.1
